split memory-schema.ts into storage, retrieval, processing, and lifecycle schemas
